bf16 not correct

+++ csrc/utils.cuh
@@ -126,7 +126,7 @@ __device__ __forceinline__ uint32_t iterator_packed_tensor(const uint32_t* ptr, 
     int second = end_bits / 32;
     start_bits = start_bits % 32;
     end_bits = end_bits % 32;
-    uint32_t v = (ptr[first] >> (start_bits)) & ((1 << WBITS) - 1);
+    uint32_t v = (ptr[first] >> (start_bits)) & (uint32_t(1 << WBITS) - 1);
     if (first == second || end_bits == 0) {
       return v;
     } else {
